requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
alembic==1.14.1
python-multipart==0.0.20
httpx==0.28.1
python-dotenv==1.1.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.3
